requirements.txt
Django==1.4.2
dj-database-url==0.2.1
django-social-auth==0.7.6
gunicorn==0.14.6
httplib2==0.7.6
ipython==0.13
oauth2==1.5.211
psycopg2==2.4.5
pygithub3==0.5
python-openid==2.2.5
requests==0.14.0
wsgiref==0.1.2
